:+1: - Add positibility to join a offers
