+++ templates/show_offer/index.html.twig
@@ -34,8 +34,13 @@
 
                         <span class="text-sm px-5 text-gray-500 dark:text-gray-400 text-center">L'équipe {{ offer.companyId.name }} aimerait vraiment t'intégrer dans notre équipe ! Rejoins-nous <span class="text-sm  italic font-bold text-gray-200"> {{ user.firstName }}</span></span>
                         <div class="flex mt-4 space-x-3 md:mt-6">
-                            <a href="#"
-                               class="inline-flex items-center py-2 px-4 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">Postuler</a>
+                            {% if(userParticipate) %}
+                                <a href="{{ path('app_participate_offer', {'id': offer.id}) }}"
+                                    class="inline-flex items-center py-2 px-4 text-sm font-medium text-center text-white bg-red-700 rounded-lg hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-800">Annuler ma candidature</a>
+                            {% else %}
+                                <a href="{{ path('app_participate_offer', {'id': offer.id}) }}"
+                                    class="inline-flex items-center py-2 px-4 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">Postuler</a>
+                            {% endif %}
                             <a href="{{ path('app_show_company', {'id': offer.companyId.id}) }}" class="inline-flex items-center py-2 px-4 text-sm font-medium text-center text-gray-900 bg-white rounded-lg border border-gray-300 hover:bg-gray-100 focus:ring-4 focus:outline-none focus:ring-gray-200 dark:bg-gray-800 dark:text-white dark:border-gray-600 dark:hover:bg-gray-700 dark:hover:border-gray-700 dark:focus:ring-gray-700">Voir toutes les offres de {{ offer.companyId.name }}</a>
                         </div>
                     </div>
